@@ -16,7 +16,6 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"runtime/debug"
@@ -126,7 +125,7 @@ func main() {
 	if oldConfigStyle == nil && (flagConfigFile != "") {
 		configFile, err := os.ReadFile(flagConfigFile)
 		if err != nil {
-			errExit("error reading config file '%s': %v", flagConfigFile, err)
+			errExit("error reading config file '%s': %v\n", flagConfigFile, err)
 		}
 		var oldConfig oldConfiguration
 		oldErr := yaml.UnmarshalStrict(configFile, &oldConfig)
@@ -143,7 +142,7 @@ func main() {
 			t := true
 			oldConfigStyle = &t
 		} else if oldErr != nil && newErr != nil {
-			errExit("error parsing configuration style as old version or new version: %v", err)
+			errExit("error parsing configuration style as old version or new version: %v\n", err)
 		}
 		// Else we fall through, and we still don't know, so we need to infer it from flags.
 	}
@@ -182,30 +181,30 @@ func main() {
 	if !*oldConfigStyle {
 		// We simply read the configuration from disk.
 		if flagConfigFile != "" {
-			buf, err := ioutil.ReadFile(flagConfigFile)
+			buf, err := os.ReadFile(flagConfigFile)
 			if err != nil {
-				errExit("error reading config file '%s': %v", flagConfigFile, err)
+				errExit("error reading config file '%s': %v\n", flagConfigFile, err)
 			}
 			err = yaml.Unmarshal(buf, &opts)
 			if err != nil {
-				errExit("error parsing'%s' as YAML: %v", flagConfigFile, err)
+				errExit("error parsing'%s' as YAML: %v\n", flagConfigFile, err)
 			}
 		}
 		var err error
 		opts, err = updateConfigFromFlags(opts)
 		if err != nil {
-			errExit("error processing flags: %v", err)
+			errExit("error processing flags: %v\n", err)
 		}
 	} else {
 		var oldConfig oldConfiguration
 		if flagConfigFile != "" {
-			buf, err := ioutil.ReadFile(flagConfigFile)
+			buf, err := os.ReadFile(flagConfigFile)
 			if err != nil {
-				errExit("error reading config file '%s': %v", flagConfigFile, err)
+				errExit("error reading config file '%s': %v\n", flagConfigFile, err)
 			}
 			err = yaml.Unmarshal(buf, &oldConfig)
 			if err != nil {
-				errExit("error parsing'%s' as YAML: %v", flagConfigFile, err)
+				errExit("error parsing'%s' as YAML: %v\n", flagConfigFile, err)
 			}
 		}
 		opts = newConfigFromOldConfig(oldConfig)
@@ -217,14 +216,14 @@ func main() {
 
 	// Now, ensure that the config options are valid.
 	if err := opts.Validate(); err != nil {
-		errExit("configuration error: %v", err)
+		errExit("configuration error: %v\n", err)
 	}
 
 	// If the user asked to output configuration, output it to stdout and exit
 	if flagOutputConfig {
 		buf, err := yaml.Marshal(opts)
 		if err != nil {
-			errExit("error YAML marshaling configuration: %v", err)
+			errExit("error YAML marshaling configuration: %v\n", err)
 		}
 		fmt.Print(string(buf))
 		return
@@ -241,9 +240,9 @@ func main() {
 	}
 
 	if opts.OutputFile != "" {
-		err = ioutil.WriteFile(opts.OutputFile, []byte(code), 0644)
+		err = os.WriteFile(opts.OutputFile, []byte(code), 0644)
 		if err != nil {
-			errExit("error writing generated code to file: %s", err)
+			errExit("error writing generated code to file: %s\n", err)
 		}
 	} else {
 		fmt.Print(code)
@@ -257,7 +256,7 @@ func loadTemplateOverrides(templatesDir string) (map[string]string, error) {
 		return templates, nil
 	}
 
-	files, err := ioutil.ReadDir(templatesDir)
+	files, err := os.ReadDir(templatesDir)
 	if err != nil {
 		return nil, err
 	}
@@ -276,7 +275,7 @@ func loadTemplateOverrides(templatesDir string) (map[string]string, error) {
 			}
 			continue
 		}
-		data, err := ioutil.ReadFile(path.Join(templatesDir, f.Name()))
+		data, err := os.ReadFile(path.Join(templatesDir, f.Name()))
 		if err != nil {
 			return nil, err
 		}
