@@ -4,7 +4,7 @@ import (
 	"context"
 	_ "embed"
 	"errors"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -96,7 +96,7 @@ func TestOapiRequestValidatorWithOptionsMultiError(t *testing.T) {
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=50")
 		assert.Equal(t, http.StatusBadRequest, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id2\"")
 			assert.Contains(t, string(body), "value is required but missing")
@@ -110,7 +110,7 @@ func TestOapiRequestValidatorWithOptionsMultiError(t *testing.T) {
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource")
 		assert.Equal(t, http.StatusBadRequest, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
 			assert.Contains(t, string(body), "value is required but missing")
@@ -126,7 +126,7 @@ func TestOapiRequestValidatorWithOptionsMultiError(t *testing.T) {
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=500")
 		assert.Equal(t, http.StatusBadRequest, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
 			assert.Contains(t, string(body), "number must be at most 100")
@@ -142,10 +142,10 @@ func TestOapiRequestValidatorWithOptionsMultiError(t *testing.T) {
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=abc&id2=1")
 		assert.Equal(t, http.StatusBadRequest, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
-			assert.Contains(t, string(body), "parsing \"abc\": invalid syntax")
+			assert.Contains(t, string(body), "value abc: an invalid integer: invalid syntax")
 			assert.Contains(t, string(body), "parameter \"id2\"")
 			assert.Contains(t, string(body), "number must be at least 10")
 		}
@@ -198,7 +198,7 @@ func TestOapiRequestValidatorWithOptionsMultiErrorAndCustomHandler(t *testing.T)
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=50")
 		assert.Equal(t, http.StatusTeapot, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id2\"")
 			assert.Contains(t, string(body), "value is required but missing")
@@ -212,7 +212,7 @@ func TestOapiRequestValidatorWithOptionsMultiErrorAndCustomHandler(t *testing.T)
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource")
 		assert.Equal(t, http.StatusTeapot, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
 			assert.Contains(t, string(body), "value is required but missing")
@@ -228,7 +228,7 @@ func TestOapiRequestValidatorWithOptionsMultiErrorAndCustomHandler(t *testing.T)
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=500")
 		assert.Equal(t, http.StatusTeapot, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
 			assert.Contains(t, string(body), "number must be at most 100")
@@ -244,10 +244,10 @@ func TestOapiRequestValidatorWithOptionsMultiErrorAndCustomHandler(t *testing.T)
 	{
 		rec := doGet(t, r, "http://deepmap.ai/multiparamresource?id=abc&id2=1")
 		assert.Equal(t, http.StatusTeapot, rec.Code)
-		body, err := ioutil.ReadAll(rec.Body)
+		body, err := io.ReadAll(rec.Body)
 		if assert.NoError(t, err) {
 			assert.Contains(t, string(body), "parameter \"id\"")
-			assert.Contains(t, string(body), "parsing \"abc\": invalid syntax")
+			assert.Contains(t, string(body), "value abc: an invalid integer: invalid syntax")
 			assert.Contains(t, string(body), "parameter \"id2\"")
 			assert.Contains(t, string(body), "number must be at least 10")
 		}
