@@ -1,3 +1,3 @@
 package client
 
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen --package=client -o client.gen.go client.yaml
+//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen --package=client --generate=client,types -o client.gen.go client.yaml
